BUG: xlwt crashes when asked to write a float32 --> convert to float 64 (fix issue #857)
Former-commit-id: 03b22e37099caf457bc1390f41c7cf833b8a45d4 [formerly 3392c23b86e2249adf764a7d432e3c83df1460bf] [formerly 00b9309fca8808bd35d2bcd53f2571d9ea63eac8 [formerly 985c83736592db31164e3c27a6d934d723678f32 [formerly 791ddd2c31dc143ca3e247cfee2503a0f8568b7a]]]
Former-commit-id: 03b22e37099caf457bc1390f41c7cf833b8a45d4 [formerly 3392c23b86e2249adf764a7d432e3c83df1460bf]
Former-commit-id: a1fc845b873d1b55cc9b2a63dc0cccf18419a02e [formerly 4282283eee9b588d455750b99a44683fa3316693 [formerly 85b351f84cf524fcb1d805338c50a1a92c37576f]]
Former-commit-id: 6a5fb6421d0d179883ef334f3bf63271f52e49bb
Former-commit-id: 3b39068fc61e935ad6c93e8ccb84a235204c1a4a

+++ scripts/sct_process_segmentation.py
@@ -749,7 +749,7 @@ def save_results(fname_output, output_type, overwrite, fname_data, metric_name, 
         sh.write(row_index, 3, vertebral_levels_field)
         sh.write(row_index, 4, slices_of_interest_field)
         sh.write(row_index, 5, os.path.abspath(fname_data))
-        sh.write(row_index, 6, mean)
+        sh.write(row_index, 6, float(mean))
         sh.write(row_index, 7, str(std))
 
         book.save(fname_output + '.' + output_type)
